Finished the basic Champion Db, and started the webpage, with components and routing
The champion Db now is able to take data but I need to rework the classes to take data and edit them. The webpage is now serveable with a component for the header and footer with a middle for Champions. Basic routing has been set up for the component, and can now request data from my API for champions using the client. The data isnt being displayed yet but it can be seen with console.log.