Added abilty to select level and health percentage for champion two and other functions
Added hide functions to the two lower parts but now three since I split off the selection of champion two. Added health percentage in the client and a API function in the API to set the health percentage. Added the function for setting champion 2. Added the needed services for champion 2. Remove a unneeded interface and edited the player interface.